refactor(database): Change HabitLog ID types to Long and update HabitLogDao
Updates the `HabitLog` entity and `HabitLogDao` to use `Long` for the primary key (`logId`) and foreign key (`habitId`) instead of `Int`. This aligns with the `Habit` entity's ID type and provides a wider range for auto-generated IDs.

- Updates the `logId` and `habitId` types in the `HabitLog` data class to `Long`.
- Modifies `HabitLogDao` methods (`getHabitLogById`, `getHabitLogsByHabitId`, `getLongestStreakForHabit`) to accept or return `Long` for the relevant IDs.
- Adjusts instrumented tests in `HabitLogDaoTest` to reflect the change in ID types and handle inserting and retrieving `Long` IDs correctly.
- Disables "resolveExternalAnnotations" in gradle.xml to address potential build issues.
- Removes an unnecessary `HabitLogDaoTest` run configuration from `deploymentTargetSelector.xml`.
- Removes an unused `kotlinx.coroutines.delay` import from `HabitDaoTest.kt`.

+++ app/src/main/java/io/track/habit/data/database/dao/HabitLogDao.kt
@@ -11,13 +11,13 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface HabitLogDao {
     @Query("SELECT * FROM habit_logs WHERE logId = :logId")
-    suspend fun getHabitLogById(logId: Int): HabitLog?
+    suspend fun getHabitLogById(logId: Long): HabitLog?
 
     @Query("SELECT * FROM habit_logs WHERE habitId = :habitId")
-    fun getHabitLogsByHabitId(habitId: Int): Flow<List<HabitLog>>
+    fun getHabitLogsByHabitId(habitId: Long): Flow<List<HabitLog>>
 
     @Query("SELECT * FROM habit_logs WHERE habitId = :habitId ORDER BY streakDuration DESC LIMIT 1")
-    fun getLongestStreakForHabit(habitId: Int): Flow<HabitLog?>
+    fun getLongestStreakForHabit(habitId: Long): Flow<HabitLog?>
 
     @Insert
     suspend fun insertHabitLog(habitLog: HabitLog)
